+++ qatrack/faults/tests/tests.py
@@ -256,7 +256,10 @@ class TestCRUDFault(TestCase):
         self.list_url = reverse("fault_list")
         self.ajax_url = reverse("fault_create_ajax")
         self.user = User.objects.create_superuser("faultuser", "a@b.com", "password")
+        self.group = qa_utils.create_group()
+        self.user.groups.add(self.group)
         self.client.force_login(self.user)
+        self.utc = qa_utils.create_unit_test_collection(unit=self.unit)
 
     def test_load_create_page(self):
         """Initial page load should work ok"""
@@ -459,7 +462,7 @@ class TestCRUDFault(TestCase):
 
         FaultType.objects.create(code="fault type")
 
-        fault = utils.create_fault()
+        fault = utils.create_fault(unit=self.unit)
 
         edit_url = reverse("fault_edit", kwargs={'pk': fault.pk})
         se = sl_utils.create_service_event()
@@ -473,7 +476,7 @@ class TestCRUDFault(TestCase):
 
         FaultType.objects.create(code="fault type")
 
-        fault = utils.create_fault()
+        fault = utils.create_fault(unit=self.unit)
         ft2 = utils.create_fault_type()
         assert fault.modality is None
 
@@ -505,7 +508,7 @@ class TestCRUDFault(TestCase):
 
         FaultType.objects.create(code="fault type")
 
-        fault = utils.create_fault()
+        fault = utils.create_fault(unit=self.unit)
         test_file = SimpleUploadedFile("TESTRUNNER_test_file.json", b"{}")
         attach = Attachment.objects.create(fault=fault, attachment=test_file, created_by=self.user)
         assert fault.attachment_set.count() == 1
@@ -541,7 +544,7 @@ class TestCRUDFault(TestCase):
 
         ft1 = utils.create_fault_type()
         ft2 = utils.create_fault_type()
-        fault = utils.create_fault(fault_type=[ft1, ft2])
+        fault = utils.create_fault(unit=self.unit, fault_type=[ft1, ft2])
         assert fault.modality is None
 
         modality = u_models.Modality.objects.create(name="modality")
@@ -575,7 +578,7 @@ class TestCRUDFault(TestCase):
 
         FaultType.objects.create(code="fault type")
 
-        fault = utils.create_fault()
+        fault = utils.create_fault(unit=self.unit)
         assert fault.modality is None
 
         edit_url = reverse("fault_edit", kwargs={'pk': fault.pk})
@@ -599,7 +602,7 @@ class TestCRUDFault(TestCase):
 
         FaultType.objects.create(code="fault type")
 
-        fault = utils.create_fault(user=self.user)
+        fault = utils.create_fault(unit=self.unit, user=self.user)
         rev_group = utils.create_fault_review_group()
         self.user.groups.add(rev_group.group)
         assert fault.modality is None
@@ -635,7 +638,7 @@ class TestCRUDFault(TestCase):
 
         FaultType.objects.create(code="fault type")
 
-        fault = utils.create_fault(user=self.user)
+        fault = utils.create_fault(unit=self.unit, user=self.user)
         rev_group = utils.create_fault_review_group()
         self.user.groups.add(rev_group.group)
         assert fault.modality is None
@@ -668,7 +671,7 @@ class TestCRUDFault(TestCase):
 
         FaultType.objects.create(code="fault type")
 
-        fault = utils.create_fault(user=self.user)
+        fault = utils.create_fault(unit=self.unit, user=self.user)
         rev_group = utils.create_fault_review_group()
         utils.create_fault_review(fault=fault, review_group=rev_group, reviewed_by=self.user)
         self.user.groups.add(rev_group.group)
@@ -705,7 +708,7 @@ class TestCRUDFault(TestCase):
 
         FaultType.objects.create(code="fault type")
 
-        fault = utils.create_fault(user=self.user)
+        fault = utils.create_fault(unit=self.unit, user=self.user)
         rev_group = utils.create_fault_review_group()
         utils.create_fault_review(fault=fault, review_group=rev_group, reviewed_by=self.user)
         self.user.groups.add(rev_group.group)
@@ -742,7 +745,7 @@ class TestCRUDFault(TestCase):
 
         FaultType.objects.create(code="fault type")
 
-        fault = utils.create_fault(user=self.user)
+        fault = utils.create_fault(unit=self.unit, user=self.user)
         rev_group = utils.create_fault_review_group(required=False)
         utils.create_fault_review(fault=fault, review_group=rev_group, reviewed_by=self.user)
         self.user.groups.add(rev_group.group)
@@ -813,7 +816,7 @@ class TestCRUDFault(TestCase):
     def test_valid_delete(self):
         """Test that deleting a fault works"""
 
-        fault = utils.create_fault()
+        fault = utils.create_fault(unit=self.unit)
         delete_url = reverse("fault_delete", kwargs={'pk': fault.pk})
         se = sl_utils.create_service_event()
         fault.related_service_events.add(se)
@@ -823,7 +826,7 @@ class TestCRUDFault(TestCase):
         assert Fault.objects.count() == 0
 
     def test_review_fault(self):
-        fault = utils.create_fault()
+        fault = utils.create_fault(unit=self.unit)
         review_url = reverse("fault_review", kwargs={'pk': fault.pk})
         resp = self.client.post(review_url)
         assert resp.status_code == 302
@@ -838,12 +841,12 @@ class TestCRUDFault(TestCase):
         assert fault.faultreviewinstance_set.first() is None
 
     def test_review_bulk(self):
-        group = qa_utils.create_group("group")
+        group = qa_utils.create_group("group2")
         self.user.groups.add(group)
         frg = FaultReviewGroup.objects.create(group=group)
 
-        f1 = utils.create_fault()
-        f2 = utils.create_fault()
+        f1 = utils.create_fault(unit=self.unit)
+        f2 = utils.create_fault(unit=self.unit)
         review_url = reverse("fault_bulk_review")
         data = {'faults': [f1.pk, f2.pk]}
         resp = self.client.post(review_url, data=data)
@@ -881,12 +884,54 @@ class TestFaultTypeAutocomplete(TestCase):
         expected = {'id': '%snew' % forms.NEW_FAULT_TYPE_MARKER, 'text': "*new*", 'code': 'new', 'description': ''}
         assert results[0] == expected
 
+    def test_query_doesnt_exist_no_add_permission(self):
+        """If query doesn't match any fault types, and user does not have fault
+        add permissions then an empty list should be returned"""
+
+        for i in range(3):
+            FaultType.objects.create(code="ft %d" % i)
+
+        user = User.objects.create_user("faultuser_non_super", email="a@b.com", password="password")
+        self.client.force_login(user)
+        results = self.client.get(self.url, {'q': 'new'}).json()['results']
+        assert len(results) == 0
+
+    def test_query_doesnt_exist_with_add_permission(self):
+        """If query doesn't match any fault types, and user does have fault
+        type add permissions then an empty list should be returned"""
+
+        for i in range(3):
+            FaultType.objects.create(code="ft %d" % i)
+
+        user = User.objects.create_user("faultuser_non_super", email="a@b.com", password="password")
+        permission = Permission.objects.get(codename="add_faulttype")
+        user.user_permissions.add(permission)
+        self.client.force_login(user)
+        results = self.client.get(self.url, {'q': 'new'}).json()['results']
+        assert len(results) == 1
+        expected = {'id': '%snew' % forms.NEW_FAULT_TYPE_MARKER, 'text': "*new*", 'code': 'new', 'description': ''}
+        assert results[0] == expected
+
     def test_query_exact_match_only(self):
         """If query matches a fault types exactly, and is not a match for any
         others, then only the exact match should be returned"""
 
         fts = [FaultType.objects.create(code="ft %d" % i) for i in range(3)]
         results = self.client.get(self.url, {'q': 'ft 2'}).json()['results']
+        expected = [{
+            'id': fts[2].code,
+            'text': "ft 2",
+            'code': fts[2].code,
+            'description': ''
+        }]
+        assert results == expected
+
+    def test_query_exact_case_insensitive_match(self):
+        """If query has a case insensitive match, and is not a match for any
+        others, then only the match should be returned"""
+
+        fts = [FaultType.objects.create(code="ft %d" % i) for i in range(3)]
+        results = self.client.get(self.url, {'q': 'FT 2'}).json()['results']
         expected = [{
             'id': fts[2].code,
             'text': "ft 2",
@@ -910,6 +955,22 @@ class TestFaultTypeAutocomplete(TestCase):
             {'id': "ft 213", 'text': "ft 213", 'code': "ft 213", 'description': ''},
         ]
 
+    def test_query_exact_match_with_whitespace_plus_others(self):
+        """If query matches a fault types but has surrounding whitespace, plus
+        matches others, all should be returned, but the exact match should be
+        first"""
+
+        FaultType.objects.create(code="ft 212", description="description"),
+        FaultType.objects.create(code="ft 21"),
+        FaultType.objects.create(code="ft 213"),
+
+        results = self.client.get(self.url, {'q': ' ft 21 '}).json()['results']
+        assert results == [
+            {'id': "ft 21", 'text': "ft 21", 'code': "ft 21", 'description': ''},
+            {'id': "ft 212", 'text': "ft 212: description", 'code': "ft 212", 'description': 'description'},
+            {'id': "ft 213", 'text': "ft 213", 'code': "ft 213", 'description': ''},
+        ]
+
 
 class TestFaultDetails(TestCase):
 
@@ -922,7 +983,7 @@ class TestFaultDetails(TestCase):
     def test_load_page(self):
         """Initial page load should work ok"""
         fault_type = FaultType.objects.create(code="ABC", slug="abc")
-        fault = utils.create_fault(fault_type=fault_type)
+        fault = utils.create_fault(unit=self.unit, fault_type=fault_type)
         url = reverse("fault_details", kwargs={'pk': fault.pk})
         resp = self.client.get(url)
         assert resp.status_code == 200
@@ -935,9 +996,9 @@ class TestFaultDetails(TestCase):
         ft1 = FaultType.objects.create(code="A", slug="a")
         ft2 = FaultType.objects.create(code="B", slug="b")
         ft3 = FaultType.objects.create(code="C", slug="c")
-        f1 = utils.create_fault(fault_type=[ft1, ft2, ft3])
-        f2 = utils.create_fault(fault_type=[ft2])
-        f3 = utils.create_fault(fault_type=[ft3])
+        f1 = utils.create_fault(unit=self.unit, fault_type=[ft1, ft2, ft3])
+        f2 = utils.create_fault(unit=self.unit, fault_type=[ft2])
+        f3 = utils.create_fault(unit=self.unit, fault_type=[ft3])
 
         from django.test import RequestFactory
 
@@ -961,7 +1022,7 @@ class TestFaultsByUnit(TestCase):
 
     def test_load_page(self):
         """Initial page load should work ok"""
-        fault = utils.create_fault()
+        fault = utils.create_fault(unit=self.unit)
         url = reverse("fault_list_by_unit", kwargs={'unit_number': fault.unit.number})
         resp = self.client.get(url, {})
         assert resp.status_code == 200
@@ -970,8 +1031,8 @@ class TestFaultsByUnit(TestCase):
 
     def test_load_result_set(self):
         """Calling via ajax should return a single object in the queryset"""
-        fault1 = utils.create_fault()
-        u2 = u_utils.create_unit()
+        fault1 = utils.create_fault(unit=self.unit)
+        u2 = qa_utils.create_unit()
         utils.create_fault(unit=u2)
         url = reverse("fault_list_by_unit", kwargs={'unit_number': fault1.unit.number})
         resp = self.client.get(url, {}, content_type='application/json', HTTP_X_REQUESTED_WITH='XMLHttpRequest')